+++ neoforge/src/main/java/io/github/xienaoban/biologydictionary/neoforge/BiologyDictionaryNeoForgeServer.java
@@ -10,18 +10,12 @@ import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.api.distmarker.OnlyIn;
-import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.common.Mod;
-import net.neoforged.fml.event.lifecycle.FMLDedicatedServerSetupEvent;
 
 @OnlyIn(Dist.DEDICATED_SERVER)
 @Mod(value = Lang.BIOLOGY_DICTIONARY, dist = Dist.DEDICATED_SERVER)
 public class BiologyDictionaryNeoForgeServer {
-    public BiologyDictionaryNeoForgeServer(IEventBus modBus) {
-        modBus.addListener(BiologyDictionaryNeoForgeServer::initServer);
-    }
-
-    private static void initServer(FMLDedicatedServerSetupEvent event) {
+    public BiologyDictionaryNeoForgeServer() {
         PacketPayloads.registerBuiltIn(new PacketPayloads.Registrar() {
             @Override
             public <T extends Packet> void register(Class<T> packetClass, Packet.Factory<T> factory) {